Refactor EmployeeDto to Employee and update related usage
The EmployeeDto object was deleted, and its usage replaced with the Employee object. This change is propagated through the PaginatedList return type of GetEmployeesWithPagination, which is used by the Get method in EmployeeController and by the GetEmployeesWithPagination handler. Two new response and request objects are introduced in a new project: CleanArchitecture.Contracts, which is now referenced in the WebApi project.